tools/invoice-automation/requirements.txt
# Invoice Automation Tool
playwright==1.40.0
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
imaplib2==3.6
email-validator==2.1.0
